tmd，不同的平台，不同的毛病

+++ start.py
@@ -151,21 +151,22 @@ def check_and_install_require_module():
 
 def call_playsound(sound):
     global played_sound_count
-    if not os.path.exists(tmp_folder):
-        os.mkdir(tmp_folder)
-    tmp_sound_path = os.path.join(tmp_folder, "{0}{1}".format(str(uuid4()), os.path.splitext(sound)[-1]))
-    shutil.copyfile(sound, tmp_sound_path)
 
-    count_lock.acquire()
-    played_sound_count += 1
-    count_lock.release()
+    if sys_platform == "Darwin":
+        if not os.path.exists(tmp_folder):
+            os.mkdir(tmp_folder)
+        tmp_sound_path = os.path.join(tmp_folder, "{0}{1}".format(str(uuid4()), os.path.splitext(sound)[-1]))
+        shutil.copyfile(sound, tmp_sound_path)
 
-    playsound(tmp_sound_path)
+        count_lock.acquire()
+        played_sound_count += 1
+        count_lock.release()
 
-    # 😮‍💨，等5秒来等待文件被其他程序释放
-    time.sleep(5)
-    
-    os.remove(tmp_sound_path)
+        playsound(tmp_sound_path)
+        os.remove(tmp_sound_path)
+
+    elif sys_platform == "Windows":
+        playsound(sound)
 
 
 def del_tmp_files():
